Major update. footnote support. many fixes.
The :footnote kind of tiddler is now supported. At the same time, the
support for :note, :tip, :warning, :caution and :important is at last
functional! These kind of tiddlers are now collectively called
'immediate' and the sample wiki has been increased to document it and
show some examples.

The support for :appendix should be good too.

Other fixes include the treatment of hyperlinks and camelcase words.