Support for wiki links with label within tables
Wiki links with a pipe inside were making havoc. But links to immediate
tiddlers are still a mess (they would require a lot of work but these
should not be there in the first place so this will be left as is).

Also support for non-breaking space entity have been added. These space
cannot be easyly replaced by character equivalence so this was a needed
fix.