Separate tiddlybook from provided sample project
Usage has shown that within real projects, the current setup was
not optimal, as command were mixed with generated files. So this
was fixed and this fix was also the solution to make a clean
separation between the tiddlybook project and its accompanying
sample project.

Now the sample project is entirely hold within a new zip archive.

Once desarchived, the sample project is a clean project which will
make it easier to use it for starting a new real project! In
particular, its directories is really what you must clone and its
updated Makefile is using it.

The README file has been updated to take this change into account.

The wiki-use.how-to has alos been udated to take into account some
externalities which required it (xmlto and required packages on
debian 9).